feat : ($RacingControllerImpl, $RacingController) RacingController 클래스를 인터페이스로 변경하고 구현하는 클래스로 RacingControllerImpl를 추가
