+++ src/main/java/racingcar/Application.java
@@ -1,13 +1,13 @@
 package racingcar;
 
 import racingcar.configuration.AppConfig;
-import racingcar.controller.RacingController;
+import racingcar.controller.RacingControllerImpl;
 
 public class Application {
     public static void main(String[] args) {
 
         AppConfig config = new AppConfig();
-        RacingController controller = config.getCarRacingController();
+        RacingControllerImpl controller = config.getCarRacingController();
         controller.run();
 
     }
